jog and move position
wrote jog code and started move position.

Gordon and I did some debugging. The jog function is fresh, but the move position needs a bit of work still. We found that move position functions in accelstepper are affected by serial monitor and other blocking functions, as well as the direction of set speed can cause the carriage to creep in the opposite direction.

Next steps, continue work on move position. Once this is done, the MVP hole drilling code can be written, and from there the fun features like sd card and screen can be added if there is time